some test changes re: feather (#104)

+++ tests/testthat/test_identical.R
@@ -14,11 +14,13 @@ test_that("Data identical (R formats)", {
     expect_equivalent(import(export(mtcars, "mtcars.R")), mtcars)
     expect_equivalent(import(export(mtcars, "mtcars.RData")), mtcars)
     expect_equivalent(import(export(mtcars, "mtcars.R", format = "dump")), mtcars)
-    expect_equivalent(import(export(mtcars, "mtcars.feather")), mtcars)
+    if (requireNamespace("feather")) {
+        expect_equivalent(import(export(mtcars, "mtcars.feather")), mtcars)
+        unlink("mtcars.feather")
+    }
     unlink("mtcars.rds")
     unlink("mtcars.R")
     unlink("mtcars.RData")
-    unlink("mtcars.feather")
 })
 
 test_that("Data identical (haven formats)", {
